Create a simple homepage
The general gist is that each elm file in `src/pages` is compiled into a
static html file and placed in the `static` directory. The web server
then serves up these static files depending on the request.

For example, if a GET request is made to "/", then the static html file
/static/home.html is returned as a response.

For convenience the commands to install elm and haskell dependencies,
generate the static html files, and build the server have been added
to the executable script `install.sh`. The script `run.sh` runs the
`install.sh` script and then runs the server.

**Note:** The path to the server executable in `run.sh` is hardcoded
and assumes the use of a 64 bit system.