Rename "crossorigin" property to "crossOrigin"
"crossorigin" is the name of the attribute, whereas "crossOrigin"
is the name of the property. There is not a 1:1 mapping between
the HTML attribute and the DOM property.

This caused chromium not to load the bootstrap resource due to CORS
being disabled thus preventing the checking of the resources integrity.