Add support for Let's Encrypt SSL Certificates
The idea is that the certificate for the domain is retrieved
automatically and then update automatically via a cron job.

At the moment the systemd services responsible for getting the certs
don't seem to work initially and need to be run manually. The nginx
service therefore fails because there are no certs.